widget: Fix text domain loading too early
Fixes text domain loading too early when initiating widget class.
The widget class itself was instantiated during module initialization while registering services. This causes translation to be loaded too early as the widget constructor was called. To mitigate this, widgets should be self contained units not instantiated during boot up but passed as a class to the register function.